brokai-lead-intel/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
httpx==0.27.0
beautifulsoup4==4.12.3
duckduckgo-search==6.2.4
langchain==0.2.6
langgraph==0.1.19
langchain-google-genai==1.0.6
langchain-groq==0.1.9
python-multipart==0.0.9
